README.md
This repo contains the code for NL2SQL paper titled "Data Agnostic RoBERTa-based Natural Language to SQL Query Generation". The link to which is: https://arxiv.org/abs/2010.05243

Our code was based on the previous works introduced in the paper "Content Enhanced BERT-based Text-to-SQL Generation", the code to which can be found here. Feel free to experiment with the models. Also these codes are developed on top of the following github repo https://github.com/DebadityaPal/RoBERTa-NL2SQL.